Moved the commponents in the assets folder to the componet folder
