+++ src/app/customizer/LoremIpsumGenerator/page.js
@@ -2,7 +2,7 @@
 import React, { useState } from "react";
 import ContentCopyIcon from "@mui/icons-material/ContentCopy";
 import CloseIcon from "@mui/icons-material/Close";
-import { NavBar } from "@/app/assets/navbar";
+import { NavBar } from "@/app/components/navbar";
 
 export default function ButtonCustomizer() {
   const [paragraphs, setParagraphs] = useState(1);
