+++ src/app/customizer/colorPicker/page.js
@@ -3,7 +3,7 @@ import React, { useState, useEffect } from "react";
 import toast, { Toaster } from "react-hot-toast";
 import ContentCopyIcon from "@mui/icons-material/ContentCopy";
 import { HexColorPicker } from "react-colorful";
-import { NavBar } from "@/app/assets/navbar";
+import { NavBar } from "@/app/components/navbar";
 
 export default function colorPicker() {
   const [hexColor, setHexColor] = useState("#ffffff");
